some correction for float

+++ 0x1E-search_algorithms/1-binary.c
@@ -24,7 +24,7 @@ int binary_search(int *array, size_t size, int value)
 		for (i = left; i < right; i++)
 			printf("%d, ", array[i]);
 		printf("%lu\n", i);
-		mid = left + (right - left) / 2;
+		mid = (int)(left + (right - left) / 2);
 		if (array[mid] == value)
 			return (mid);
 		else if (array[mid] < value)
